renaming and moving a file
