+++ blockchain/calls/proxyActions/adapters/charteredDssProxyActionsContractWrapper.ts
@@ -1,66 +1,37 @@
-import Web3 from 'web3'
-
-import { DssProxyActions } from '../../../types/web3-v1-contracts/dss-proxy-actions'
+import { DssProxyActionsCharter } from '../../../../types/web3-v1-contracts/dss-proxy-actions-charter'
 import {
   NonPayableTransactionObject,
   PayableTransactionObject,
-} from '../../../types/web3-v1-contracts/types'
-import { ContextConnected } from '../../network'
-import { amountToWei, amountToWeiRoundDown } from '../../utils'
+} from '../../../../types/web3-v1-contracts/types'
+import { ContextConnected } from '../../../network'
+import { amountToWei, amountToWeiRoundDown } from '../../../utils'
 import { DssProxyActionsSmartContractWrapperInterface } from './DssProxyActionsSmartContractWrapperInterface'
-import { DepositAndGenerateData, OpenData, WithdrawAndPaybackData } from './proxyActions'
+import { DepositAndGenerateData, OpenData, WithdrawAndPaybackData } from '../proxyActions'
 
-export const StandardDssProxyActionsContractWrapper: DssProxyActionsSmartContractWrapperInterface = {
+export const CharteredDssProxyActionsContractWrapper: DssProxyActionsSmartContractWrapperInterface = {
+  open(context: ContextConnected, data: OpenData): NonPayableTransactionObject<string> {
+    throw 'cannot open chartered vault manually'
+  },
+
   openLockETHAndDraw(context: ContextConnected, data: OpenData): PayableTransactionObject<string> {
-    const { dssCdpManager, mcdJoinDai, mcdJug, joins, contract } = context
-    const { generateAmount, ilk } = data
-    return contract<DssProxyActions>(context.dssProxyActions).methods.openLockETHAndDraw(
-      dssCdpManager.address,
-      mcdJug.address,
-      joins[ilk],
-      mcdJoinDai.address,
-      Web3.utils.utf8ToHex(ilk),
-      amountToWei(generateAmount, 'DAI').toFixed(0),
-    )
+    throw 'cannot open chartered vault manually'
   },
 
   openLockGemAndDraw(
     context: ContextConnected,
     data: OpenData,
   ): NonPayableTransactionObject<string> {
-    const { dssProxyActions, dssCdpManager, mcdJoinDai, mcdJug, joins, contract } = context
-    const { depositAmount, generateAmount, token, ilk } = data
-    return contract<DssProxyActions>(dssProxyActions).methods.openLockGemAndDraw(
-      dssCdpManager.address,
-      mcdJug.address,
-      joins[ilk],
-      mcdJoinDai.address,
-      Web3.utils.utf8ToHex(ilk),
-      amountToWei(depositAmount, token).toFixed(0),
-      amountToWei(generateAmount, 'DAI').toFixed(0),
-      true,
-    )
-  },
-
-  open(context: ContextConnected, data: OpenData): NonPayableTransactionObject<string> {
-    const { dssProxyActions, dssCdpManager, contract } = context
-    const { ilk, proxyAddress } = data
-    return contract<DssProxyActions>(dssProxyActions).methods.open(
-      dssCdpManager.address,
-      Web3.utils.utf8ToHex(ilk),
-      proxyAddress,
-    )
+    throw 'cannot open chartered vault manually'
   },
 
   resolveContractAddress(context: ContextConnected): string {
-    return context.dssProxyActions.address
+    return context.dssProxyActionsCharter.address
   },
 
   draw(context: ContextConnected, data: DepositAndGenerateData): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActions>(context.dssProxyActions)
+      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
       .methods.draw(
-        context.dssCdpManager.address,
         context.mcdJug.address,
         context.mcdJoinDai.address,
         data.id.toString(),
@@ -73,9 +44,8 @@ export const StandardDssProxyActionsContractWrapper: DssProxyActionsSmartContrac
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActions>(context.dssProxyActions)
+      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
       .methods.freeETH(
-        context.dssCdpManager.address,
         context.joins[data.ilk],
         data.id.toString(),
         amountToWei(data.withdrawAmount, data.token).toFixed(0),
@@ -87,9 +57,8 @@ export const StandardDssProxyActionsContractWrapper: DssProxyActionsSmartContrac
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActions>(context.dssProxyActions)
+      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
       .methods.freeGem(
-        context.dssCdpManager.address,
         context.joins[data.ilk],
         data.id.toString(),
         amountToWeiRoundDown(data.withdrawAmount, data.token).toFixed(0),
@@ -98,8 +67,8 @@ export const StandardDssProxyActionsContractWrapper: DssProxyActionsSmartContrac
 
   lockETH(context: ContextConnected, data: DepositAndGenerateData): PayableTransactionObject<void> {
     return context
-      .contract<DssProxyActions>(context.dssProxyActions)
-      .methods.lockETH(context.dssCdpManager.address, context.joins[data.ilk], data.id.toString())
+      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
+      .methods.lockETH(context.joins[data.ilk], data.id.toString())
   },
 
   lockETHAndDraw(
@@ -107,9 +76,8 @@ export const StandardDssProxyActionsContractWrapper: DssProxyActionsSmartContrac
     data: DepositAndGenerateData,
   ): PayableTransactionObject<void> {
     return context
-      .contract<DssProxyActions>(context.dssProxyActions)
+      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
       .methods.lockETHAndDraw(
-        context.dssCdpManager.address,
         context.mcdJug.address,
         context.joins[data.ilk],
         context.mcdJoinDai.address,
@@ -123,13 +91,11 @@ export const StandardDssProxyActionsContractWrapper: DssProxyActionsSmartContrac
     data: DepositAndGenerateData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActions>(context.dssProxyActions)
+      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
       .methods.lockGem(
-        context.dssCdpManager.address,
         context.joins[data.ilk],
         data.id.toString(),
         amountToWei(data.depositAmount, data.token).toFixed(0),
-        true,
       )
   },
 
@@ -138,24 +104,21 @@ export const StandardDssProxyActionsContractWrapper: DssProxyActionsSmartContrac
     data: DepositAndGenerateData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActions>(context.dssProxyActions)
+      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
       .methods.lockGemAndDraw(
-        context.dssCdpManager.address,
         context.mcdJug.address,
         context.joins[data.ilk],
         context.mcdJoinDai.address,
         data.id.toString(),
         amountToWei(data.depositAmount, data.token).toFixed(0),
         amountToWei(data.generateAmount, 'DAI').toFixed(0),
-        true,
       )
   },
 
   wipe(context: ContextConnected, data: WithdrawAndPaybackData): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActions>(context.dssProxyActions)
+      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
       .methods.wipe(
-        context.dssCdpManager.address,
         context.mcdJoinDai.address,
         data.id.toString(),
         amountToWei(data.paybackAmount, 'DAI').toFixed(0),
@@ -167,12 +130,8 @@ export const StandardDssProxyActionsContractWrapper: DssProxyActionsSmartContrac
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActions>(context.dssProxyActions)
-      .methods.wipeAll(
-        context.dssCdpManager.address,
-        context.mcdJoinDai.address,
-        data.id.toString(),
-      )
+      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
+      .methods.wipeAll(context.mcdJoinDai.address, data.id.toString())
   },
 
   wipeAllAndFreeETH(
@@ -180,9 +139,8 @@ export const StandardDssProxyActionsContractWrapper: DssProxyActionsSmartContrac
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActions>(context.dssProxyActions)
+      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
       .methods.wipeAllAndFreeETH(
-        context.dssCdpManager.address,
         context.joins[data.ilk],
         context.mcdJoinDai.address,
         data.id.toString(),
@@ -195,9 +153,8 @@ export const StandardDssProxyActionsContractWrapper: DssProxyActionsSmartContrac
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActions>(context.dssProxyActions)
+      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
       .methods.wipeAllAndFreeGem(
-        context.dssCdpManager.address,
         context.joins[data.ilk],
         context.mcdJoinDai.address,
         data.id.toString(),
@@ -210,9 +167,8 @@ export const StandardDssProxyActionsContractWrapper: DssProxyActionsSmartContrac
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActions>(context.dssProxyActions)
+      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
       .methods.wipeAndFreeETH(
-        context.dssCdpManager.address,
         context.joins[data.ilk],
         context.mcdJoinDai.address,
         data.id.toString(),
@@ -226,9 +182,8 @@ export const StandardDssProxyActionsContractWrapper: DssProxyActionsSmartContrac
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActions>(context.dssProxyActions)
+      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
       .methods.wipeAndFreeGem(
-        context.dssCdpManager.address,
         context.joins[data.ilk],
         context.mcdJoinDai.address,
         data.id.toString(),
