+++ blockchain/calls/proxyActions/adapters/dssProxyActionsSmartContractWrapper.mock.ts
@@ -1,96 +1,116 @@
 import {
   NonPayableTransactionObject,
   PayableTransactionObject,
-} from '../../../types/web3-v1-contracts/types'
-import { ContextConnected } from '../../network'
+  PayableTx,
+} from '../../../../types/web3-v1-contracts/types'
+import { ContextConnected } from '../../../network'
 import { DssProxyActionsSmartContractWrapperInterface } from './DssProxyActionsSmartContractWrapperInterface'
-import { DepositAndGenerateData, OpenData, WithdrawAndPaybackData } from './proxyActions'
+import { DepositAndGenerateData, OpenData, WithdrawAndPaybackData } from '../proxyActions'
 
-export const CurveDssProxyActionsSmartContractWrapper: DssProxyActionsSmartContractWrapperInterface = {
+function createMockPayableTransactionObject<T>(name: string): PayableTransactionObject<T> {
+  return {
+    arguments: [],
+    call: (tx?: PayableTx) => {
+      return Promise.reject('not implemented')
+    },
+    encodeABI: () => {
+      const call = {
+        method: name,
+      }
+      return JSON.stringify(call)
+    },
+    estimateGas: (tx?: PayableTx) => {
+      return Promise.reject('not implemented')
+    },
+    send: (tx?: PayableTx) => {
+      throw 'not implemented'
+    },
+  }
+}
+
+export const MockDssProxyActionsSmartContractWrapper: DssProxyActionsSmartContractWrapperInterface = {
   open(context: ContextConnected, data: OpenData): NonPayableTransactionObject<string> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject<string>('open')
   },
-
   openLockETHAndDraw(context: ContextConnected, data: OpenData): PayableTransactionObject<string> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('openLockETHAndDraw')
   },
-
   openLockGemAndDraw(
     context: ContextConnected,
     data: OpenData,
   ): NonPayableTransactionObject<string> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('openLockGemAndDraw')
   },
   draw(context: ContextConnected, data: DepositAndGenerateData): NonPayableTransactionObject<void> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('draw')
   },
   freeETH(
     context: ContextConnected,
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('freeETH')
   },
   freeGem(
     context: ContextConnected,
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('freeGem')
   },
   lockETH(context: ContextConnected, data: DepositAndGenerateData): PayableTransactionObject<void> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('lockETH')
   },
   lockETHAndDraw(
     context: ContextConnected,
     data: DepositAndGenerateData,
   ): PayableTransactionObject<void> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('lockETHAndDraw')
   },
   lockGem(
     context: ContextConnected,
     data: DepositAndGenerateData,
   ): NonPayableTransactionObject<void> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('lockGem')
   },
   lockGemAndDraw(
     context: ContextConnected,
     data: DepositAndGenerateData,
   ): NonPayableTransactionObject<void> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('lockGemAndDraw')
   },
   resolveContractAddress(context: ContextConnected): string {
-    return context.dssProxyActionsCurve.address
+    return '0x-mock-dss-proxy-action-address'
   },
   wipe(context: ContextConnected, data: WithdrawAndPaybackData): NonPayableTransactionObject<void> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('wipe')
   },
   wipeAll(
     context: ContextConnected,
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('wipeAll')
   },
   wipeAllAndFreeETH(
     context: ContextConnected,
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('wipeAllAndFreeETH')
   },
   wipeAllAndFreeGem(
     context: ContextConnected,
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('wipeAllAndFreeGem')
   },
   wipeAndFreeETH(
     context: ContextConnected,
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('wipeAndFreeETH')
   },
   wipeAndFreeGem(
     context: ContextConnected,
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
-    throw 'unimplemented'
+    return createMockPayableTransactionObject('wipeAndFreeGem')
   },
 }
