+++ blockchain/calls/proxyActions/adapters/standardDssProxyActionsContractWrapper.ts
@@ -1,37 +1,66 @@
-import { DssProxyActionsCharter } from '../../../types/web3-v1-contracts/dss-proxy-actions-charter'
+import Web3 from 'web3'
+
+import { DssProxyActions } from '../../../../types/web3-v1-contracts/dss-proxy-actions'
 import {
   NonPayableTransactionObject,
   PayableTransactionObject,
-} from '../../../types/web3-v1-contracts/types'
-import { ContextConnected } from '../../network'
-import { amountToWei, amountToWeiRoundDown } from '../../utils'
+} from '../../../../types/web3-v1-contracts/types'
+import { ContextConnected } from '../../../network'
+import { amountToWei, amountToWeiRoundDown } from '../../../utils'
 import { DssProxyActionsSmartContractWrapperInterface } from './DssProxyActionsSmartContractWrapperInterface'
-import { DepositAndGenerateData, OpenData, WithdrawAndPaybackData } from './proxyActions'
+import { DepositAndGenerateData, OpenData, WithdrawAndPaybackData } from '../proxyActions'
 
-export const CharteredDssProxyActionsContractWrapper: DssProxyActionsSmartContractWrapperInterface = {
-  open(context: ContextConnected, data: OpenData): NonPayableTransactionObject<string> {
-    throw 'cannot open chartered vault manually'
-  },
-
+export const StandardDssProxyActionsContractWrapper: DssProxyActionsSmartContractWrapperInterface = {
   openLockETHAndDraw(context: ContextConnected, data: OpenData): PayableTransactionObject<string> {
-    throw 'cannot open chartered vault manually'
+    const { dssCdpManager, mcdJoinDai, mcdJug, joins, contract } = context
+    const { generateAmount, ilk } = data
+    return contract<DssProxyActions>(context.dssProxyActions).methods.openLockETHAndDraw(
+      dssCdpManager.address,
+      mcdJug.address,
+      joins[ilk],
+      mcdJoinDai.address,
+      Web3.utils.utf8ToHex(ilk),
+      amountToWei(generateAmount, 'DAI').toFixed(0),
+    )
   },
 
   openLockGemAndDraw(
     context: ContextConnected,
     data: OpenData,
   ): NonPayableTransactionObject<string> {
-    throw 'cannot open chartered vault manually'
+    const { dssProxyActions, dssCdpManager, mcdJoinDai, mcdJug, joins, contract } = context
+    const { depositAmount, generateAmount, token, ilk } = data
+    return contract<DssProxyActions>(dssProxyActions).methods.openLockGemAndDraw(
+      dssCdpManager.address,
+      mcdJug.address,
+      joins[ilk],
+      mcdJoinDai.address,
+      Web3.utils.utf8ToHex(ilk),
+      amountToWei(depositAmount, token).toFixed(0),
+      amountToWei(generateAmount, 'DAI').toFixed(0),
+      true,
+    )
+  },
+
+  open(context: ContextConnected, data: OpenData): NonPayableTransactionObject<string> {
+    const { dssProxyActions, dssCdpManager, contract } = context
+    const { ilk, proxyAddress } = data
+    return contract<DssProxyActions>(dssProxyActions).methods.open(
+      dssCdpManager.address,
+      Web3.utils.utf8ToHex(ilk),
+      proxyAddress,
+    )
   },
 
   resolveContractAddress(context: ContextConnected): string {
-    return context.dssProxyActionsCharter.address
+    return context.dssProxyActions.address
   },
 
   draw(context: ContextConnected, data: DepositAndGenerateData): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
+      .contract<DssProxyActions>(context.dssProxyActions)
       .methods.draw(
+        context.dssCdpManager.address,
         context.mcdJug.address,
         context.mcdJoinDai.address,
         data.id.toString(),
@@ -44,8 +73,9 @@ export const CharteredDssProxyActionsContractWrapper: DssProxyActionsSmartContra
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
+      .contract<DssProxyActions>(context.dssProxyActions)
       .methods.freeETH(
+        context.dssCdpManager.address,
         context.joins[data.ilk],
         data.id.toString(),
         amountToWei(data.withdrawAmount, data.token).toFixed(0),
@@ -57,8 +87,9 @@ export const CharteredDssProxyActionsContractWrapper: DssProxyActionsSmartContra
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
+      .contract<DssProxyActions>(context.dssProxyActions)
       .methods.freeGem(
+        context.dssCdpManager.address,
         context.joins[data.ilk],
         data.id.toString(),
         amountToWeiRoundDown(data.withdrawAmount, data.token).toFixed(0),
@@ -67,8 +98,8 @@ export const CharteredDssProxyActionsContractWrapper: DssProxyActionsSmartContra
 
   lockETH(context: ContextConnected, data: DepositAndGenerateData): PayableTransactionObject<void> {
     return context
-      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
-      .methods.lockETH(context.joins[data.ilk], data.id.toString())
+      .contract<DssProxyActions>(context.dssProxyActions)
+      .methods.lockETH(context.dssCdpManager.address, context.joins[data.ilk], data.id.toString())
   },
 
   lockETHAndDraw(
@@ -76,8 +107,9 @@ export const CharteredDssProxyActionsContractWrapper: DssProxyActionsSmartContra
     data: DepositAndGenerateData,
   ): PayableTransactionObject<void> {
     return context
-      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
+      .contract<DssProxyActions>(context.dssProxyActions)
       .methods.lockETHAndDraw(
+        context.dssCdpManager.address,
         context.mcdJug.address,
         context.joins[data.ilk],
         context.mcdJoinDai.address,
@@ -91,11 +123,13 @@ export const CharteredDssProxyActionsContractWrapper: DssProxyActionsSmartContra
     data: DepositAndGenerateData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
+      .contract<DssProxyActions>(context.dssProxyActions)
       .methods.lockGem(
+        context.dssCdpManager.address,
         context.joins[data.ilk],
         data.id.toString(),
         amountToWei(data.depositAmount, data.token).toFixed(0),
+        true,
       )
   },
 
@@ -104,21 +138,24 @@ export const CharteredDssProxyActionsContractWrapper: DssProxyActionsSmartContra
     data: DepositAndGenerateData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
+      .contract<DssProxyActions>(context.dssProxyActions)
       .methods.lockGemAndDraw(
+        context.dssCdpManager.address,
         context.mcdJug.address,
         context.joins[data.ilk],
         context.mcdJoinDai.address,
         data.id.toString(),
         amountToWei(data.depositAmount, data.token).toFixed(0),
         amountToWei(data.generateAmount, 'DAI').toFixed(0),
+        true,
       )
   },
 
   wipe(context: ContextConnected, data: WithdrawAndPaybackData): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
+      .contract<DssProxyActions>(context.dssProxyActions)
       .methods.wipe(
+        context.dssCdpManager.address,
         context.mcdJoinDai.address,
         data.id.toString(),
         amountToWei(data.paybackAmount, 'DAI').toFixed(0),
@@ -130,8 +167,12 @@ export const CharteredDssProxyActionsContractWrapper: DssProxyActionsSmartContra
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
-      .methods.wipeAll(context.mcdJoinDai.address, data.id.toString())
+      .contract<DssProxyActions>(context.dssProxyActions)
+      .methods.wipeAll(
+        context.dssCdpManager.address,
+        context.mcdJoinDai.address,
+        data.id.toString(),
+      )
   },
 
   wipeAllAndFreeETH(
@@ -139,8 +180,9 @@ export const CharteredDssProxyActionsContractWrapper: DssProxyActionsSmartContra
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
+      .contract<DssProxyActions>(context.dssProxyActions)
       .methods.wipeAllAndFreeETH(
+        context.dssCdpManager.address,
         context.joins[data.ilk],
         context.mcdJoinDai.address,
         data.id.toString(),
@@ -153,8 +195,9 @@ export const CharteredDssProxyActionsContractWrapper: DssProxyActionsSmartContra
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
+      .contract<DssProxyActions>(context.dssProxyActions)
       .methods.wipeAllAndFreeGem(
+        context.dssCdpManager.address,
         context.joins[data.ilk],
         context.mcdJoinDai.address,
         data.id.toString(),
@@ -167,8 +210,9 @@ export const CharteredDssProxyActionsContractWrapper: DssProxyActionsSmartContra
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
+      .contract<DssProxyActions>(context.dssProxyActions)
       .methods.wipeAndFreeETH(
+        context.dssCdpManager.address,
         context.joins[data.ilk],
         context.mcdJoinDai.address,
         data.id.toString(),
@@ -182,8 +226,9 @@ export const CharteredDssProxyActionsContractWrapper: DssProxyActionsSmartContra
     data: WithdrawAndPaybackData,
   ): NonPayableTransactionObject<void> {
     return context
-      .contract<DssProxyActionsCharter>(context.dssProxyActionsCharter)
+      .contract<DssProxyActions>(context.dssProxyActions)
       .methods.wipeAndFreeGem(
+        context.dssCdpManager.address,
         context.joins[data.ilk],
         context.mcdJoinDai.address,
         data.id.toString(),
