+++ components/AppContext.ts
@@ -34,7 +34,7 @@ import {
   OpenMultiplyData,
   ReclaimData,
   WithdrawAndPaybackData,
-  withdrawPaybackDepositGenerateLogicFactory,
+  vaultActionsLogicFactory,
 } from 'blockchain/calls/proxyActions/proxyActions'
 import { vatGem, vatIlk, vatUrns } from 'blockchain/calls/vat'
 import { resolveENSName$ } from 'blockchain/ens'
@@ -125,8 +125,8 @@ import {
 } from '../blockchain/calls/erc20'
 import { jugIlk } from '../blockchain/calls/jug'
 import { observe } from '../blockchain/calls/observe'
-import { CharteredDssProxyActionsContractWrapper } from '../blockchain/calls/proxyActions/charteredDssProxyActionsContractWrapper'
-import { StandardDssProxyActionsContractWrapper } from '../blockchain/calls/proxyActions/standardDssProxyActionsContractWrapper'
+import { CharteredDssProxyActionsContractWrapper } from '../blockchain/calls/proxyActions/adapters/charteredDssProxyActionsContractWrapper'
+import { StandardDssProxyActionsContractWrapper } from '../blockchain/calls/proxyActions/adapters/standardDssProxyActionsContractWrapper'
 import { spotIlk } from '../blockchain/calls/spot'
 import { networksById } from '../blockchain/config'
 import {
@@ -575,7 +575,7 @@ export function setupAppContext() {
         saveVaultUsingApi$,
         addGasEstimation$,
         vaultHistory$,
-        withdrawPaybackDepositGenerateLogicFactory(StandardDssProxyActionsContractWrapper),
+        vaultActionsLogicFactory(StandardDssProxyActionsContractWrapper),
         StandardBorrowManageVaultViewStateProvider,
         id,
       ),
@@ -596,7 +596,7 @@ export function setupAppContext() {
         saveVaultUsingApi$,
         addGasEstimation$,
         vaultHistory$,
-        withdrawPaybackDepositGenerateLogicFactory(CharteredDssProxyActionsContractWrapper),
+        vaultActionsLogicFactory(CharteredDssProxyActionsContractWrapper),
         // comment out above and uncomment below to test insti vault flows + UI against standard borrow vault
         // withdrawPaybackDepositGenerateLogicFactory(StandardDssProxyActionsContractWrapper),
         InstitutionalBorrowManageVaultViewStateProvider,
