+++ features/borrow/open/pipes/openVaultTransactions.ts
@@ -1,8 +1,5 @@
 import { TxStatus } from '@oasisdex/transactions'
-import {
-  OpenData,
-  withdrawPaybackDepositGenerateLogicFactory,
-} from 'blockchain/calls/proxyActions/proxyActions'
+import { OpenData, vaultActionsLogicFactory } from 'blockchain/calls/proxyActions/proxyActions'
 import { TxMetaKind } from 'blockchain/calls/txMeta'
 import { AddGasEstimationFunction, TxHelpers } from 'components/AppContext'
 import { VaultType } from 'features/generalManageVault/vaultType'
@@ -12,7 +9,7 @@ import { transactionToX } from 'helpers/form'
 import { zero } from 'helpers/zero'
 import { Observable, of } from 'rxjs'
 
-import { StandardDssProxyActionsContractWrapper } from '../../../../blockchain/calls/proxyActions/standardDssProxyActionsContractWrapper'
+import { StandardDssProxyActionsContractWrapper } from '../../../../blockchain/calls/proxyActions/adapters/standardDssProxyActionsContractWrapper'
 import { parseVaultIdFromReceiptLogs } from '../../../shared/transactions'
 import { OpenVaultChange, OpenVaultState } from './openVault'
 
@@ -87,17 +84,14 @@ export function openVault(
   change: (ch: OpenVaultChange) => void,
   { generateAmount, depositAmount, proxyAddress, ilk, account, token }: OpenVaultState,
 ) {
-  sendWithGasEstimation(
-    withdrawPaybackDepositGenerateLogicFactory(StandardDssProxyActionsContractWrapper).open,
-    {
-      kind: TxMetaKind.open,
-      generateAmount: generateAmount || zero,
-      depositAmount: depositAmount || zero,
-      proxyAddress: proxyAddress!,
-      ilk,
-      token,
-    },
-  )
+  sendWithGasEstimation(vaultActionsLogicFactory(StandardDssProxyActionsContractWrapper).open, {
+    kind: TxMetaKind.open,
+    generateAmount: generateAmount || zero,
+    depositAmount: depositAmount || zero,
+    proxyAddress: proxyAddress!,
+    ilk,
+    token,
+  })
     .pipe(
       transactionToX<OpenVaultChange, OpenData>(
         { kind: 'txWaitingForApproval' },
@@ -144,17 +138,14 @@ export function applyEstimateGas(
     const { proxyAddress, generateAmount, depositAmount, ilk, token } = state
 
     if (proxyAddress && (generateAmount || depositAmount)) {
-      return estimateGas(
-        withdrawPaybackDepositGenerateLogicFactory(StandardDssProxyActionsContractWrapper).open,
-        {
-          kind: TxMetaKind.open,
-          generateAmount: generateAmount || zero,
-          depositAmount: depositAmount || zero,
-          proxyAddress,
-          ilk,
-          token,
-        },
-      )
+      return estimateGas(vaultActionsLogicFactory(StandardDssProxyActionsContractWrapper).open, {
+        kind: TxMetaKind.open,
+        generateAmount: generateAmount || zero,
+        depositAmount: depositAmount || zero,
+        proxyAddress,
+        ilk,
+        token,
+      })
     }
 
     return undefined
