+++ helpers/mocks/manageVault.mock.ts
@@ -16,8 +16,8 @@ import { one, zero } from 'helpers/zero'
 import { Observable, of } from 'rxjs'
 import { switchMap } from 'rxjs/operators'
 
-import { withdrawPaybackDepositGenerateLogicFactory } from '../../blockchain/calls/proxyActions/proxyActions'
-import { StandardDssProxyActionsContractWrapper } from '../../blockchain/calls/proxyActions/standardDssProxyActionsContractWrapper'
+import { vaultActionsLogicFactory } from '../../blockchain/calls/proxyActions/proxyActions'
+import { StandardDssProxyActionsContractWrapper } from '../../blockchain/calls/proxyActions/adapters/standardDssProxyActionsContractWrapper'
 import { createInstiVault$, InstiVault } from '../../blockchain/instiVault'
 import { InstitutionalBorrowManageVaultViewStateProvider } from '../../features/borrow/manage/pipes/viewStateProviders/institutionalBorrowManageVaultViewStateProvider'
 import { StandardBorrowManageVaultViewStateProvider } from '../../features/borrow/manage/pipes/viewStateProviders/standardBorrowManageVaultViewStateProvider'
@@ -224,7 +224,7 @@ export function mockManageVault$(
     saveVaultType$,
     addGasEstimationMock,
     vaultHistory$,
-    withdrawPaybackDepositGenerateLogicFactory(StandardDssProxyActionsContractWrapper),
+    vaultActionsLogicFactory(StandardDssProxyActionsContractWrapper),
     StandardBorrowManageVaultViewStateProvider,
     MOCK_VAULT_ID,
   )
@@ -282,7 +282,7 @@ export function mockManageInstiVault$(
     saveVaultType$,
     addGasEstimationMock,
     vaultHistory$,
-    withdrawPaybackDepositGenerateLogicFactory(StandardDssProxyActionsContractWrapper),
+    vaultActionsLogicFactory(StandardDssProxyActionsContractWrapper),
     InstitutionalBorrowManageVaultViewStateProvider,
     MOCK_VAULT_ID,
   )
